+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 New features:
 
-  - First steps to implement lock requst mode "2". This is not supported yet, as the firmware of the bridge is not published, yet.
-  - Added methods to bride for reboot, firmware update, and for refreshing all locks.
+  - Implemented lock state mode "2".
+  - Added methods to bridge for reboot, firmware update, and for refreshing all locks.
 
 Bugfix: